policy table, added non-localized timestamp
